writing: extract rule-health verdict domain from report
